requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-crispy-forms==2.0
django-gravatar2==1.4.4
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.3
